added some NBench.PerformanceCounter integration tests - discovered that we weren't logging any metrics updated all collectors to use floating point made it possible to identify the correct NBench process instance added object pooling for PerformanceCounter

+++ src/NBench/Collection/Memory/GcTotalMemoryCollector.cs
@@ -24,7 +24,7 @@ namespace NBench.Collection.Memory
         {
         }
 
-        public override long Collect()
+        public override double Collect()
         {
             /*
              * We intentionally don't allow the garbage collector to collect anything
